Remove Manifold dependency (#9)
The [manifold](https://github.com/clj-commons/manifold/) library is a
great async abstraction layer, but not every client who wants to use
`ken` will be using manifold, and it'd be nice to not have to pull it
into their dependency closures. There's a pretty simple pattern for
doing this in Clojure, which is to test for a namespace existing at
compile-time using a macro.

Interestingly, as of 0.4.x, manifold `Deferred` values also implement
`CompletionStage`, so the generic code path in #8 would also work
without any dependency at all! Unfortunately, there's one particular
implementation wrinkle in manifold that causes very unintuitive behavior
when used with `ken/watch` in a certain way. There's a simple fix for
this using `bound-fn` wrappers, but we do need manifold-specific logic
to do so. I added a new test with extensive comments to ensure that this
doesn't regress.

Update: after further experimentation and consulting the ~ancient
scrolls~ source code, I found a way to effectively mimic what the
manifold callbacks were doing using Clojure's `Var` manipulation
directly. 🧙‍♂️